requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
torch>=2.0.0
torchaudio>=2.0.0
demucs>=4.0.0
librosa>=0.10.0
matplotlib>=3.7.0
numpy>=1.24.0
Werkzeug>=2.3.0
eventlet>=0.33.0
python-socketio>=5.8.0
gunicorn==21.2.0
waitress==3.0.0
python-engineio>=4.3.4